[update] Package publish preparations
