+++ pubspec.yaml
@@ -1,6 +1,6 @@
 name: badges
 description: A flutter package for creating badges.
-version: 0.0.6
+version: 1.0.0
 author: Daniyil Yakovlev <danny.yakovlev@gmail.com>
 homepage: https://github.com/yadaniyil/flutter_badges
